L_4/L4_prompts/chat_prompt_dynamic.py
from langchain_google_genai import ChatGoogleGenerativeAI
from langchain_core.messages import SystemMessage,HumanMessage
from langchain_core.prompts import ChatPromptTemplate
from dotenv import load_dotenv

load_dotenv()
model = ChatGoogleGenerativeAI(model="gemini-2.0-flash")

chat_prompt = ChatPromptTemplate([
    ('system','You are a helpful {domain} expert'),
    ('human','Explain in simple terms about {topic}'),
])

prompt = chat_prompt.invoke({'domain':'Computer Science','topic' : 'networking'});

print(prompt)