L_4/requirement.txt
# Langchain core
langchain
langchain-core

# OpenAI integration
langchain-openai
openai

# Anthropic integration
langchain-anthropic

# Google Gemini (PaLM) integration
langchain-google-genai
google-generativeai

# Hugging Face integration
langchain-huggingface
transformers
huggingface-hub

# Environment variable management
python-dotenv

# Machine learning utilities
numpy
scikit-learn
